add checks; improved documentation

+++ Apps/WebAPI/sdLitica.WebAPI/Controllers/v1/TimeSeriesController.cs
@@ -10,7 +10,6 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace sdLitica.WebAPI.Controllers.v1
@@ -19,8 +18,7 @@ namespace sdLitica.WebAPI.Controllers.v1
     /// This controller is used to work with user timeseries
     /// </summary>
     [Route("api/v1/timeseries")]
-    //TODO: Change auth settings here
-    [AllowAnonymous]
+    [Authorize]
     public class TimeSeriesController : BaseApiController
     {
         private readonly ITimeSeriesService _timeSeriesService;
@@ -40,12 +38,21 @@ namespace sdLitica.WebAPI.Controllers.v1
         ///    400 - There were issues with passed data, i.e. required fields missing or length constraints violated
         /// </returns>
         [HttpPost]
+        [AllowAnonymous]
         public IActionResult AddTimeSeries()
         {
             var t = _timeSeriesService.AddRandomTimeSeries();
             return Ok(t.Result);
         }
 
+        /// <summary>
+        /// This REST API handler creates a new time-series metadata object for current user
+        /// </summary>
+        /// <param name="timeSeriesModel"></param>
+        /// <returns>
+        ///    200 - Time-series was successfully created. Response payload will contain object with assigned id
+        ///    400 - There were issues with passed data, i.e. required fields missing or length constraints violated
+        /// </returns>
         [HttpPost]
         [Route("temp")]
         public IActionResult AddTimeSeries([FromBody] TimeSeriesMetadataModel timeSeriesModel) {
@@ -54,17 +61,29 @@ namespace sdLitica.WebAPI.Controllers.v1
             return Ok(new TimeSeriesMetadataModel(t.Result));
         }
 
+        /// <summary>
+        /// This REST API handler uploads a data from csv-file to the time-series given by timeSeriesMetadataId
+        /// </summary>
+        /// <param name="timeSeriesMetadataId"></param>
+        /// <param name="formFile"></param>
+        /// <returns></returns>
         [HttpPost]
         [Route("{measurementId}/data")]
-        public IActionResult UploadCsvData([FromRoute] string measurementId, [FromForm] IFormFile formFile)
+        public IActionResult UploadCsvData([FromRoute] string timeSeriesMetadataId, [FromForm] IFormFile formFile)
         {
-            // todo: check if authorized user owns this time-series, update rows and columns metadata after extraction
+            // todo: update rows and columns metadata after extraction
+            TimeSeriesMetadata timeSeriesMetadata = _timeSeriesMetadataService.GetTimeSeriesMetadata(timeSeriesMetadataId);
+            if (!(timeSeriesMetadata != null && timeSeriesMetadata.UserId.Equals(UserId)))
+            {
+                return NotFound("this user does not have timeseries given by this id");
+            }
+            string measurementId = timeSeriesMetadata.InfluxId.ToString();
             List<string> fileContent = ReadAsStringAsync(formFile).Result;
             _timeSeriesService.UploadDataFromCsv(measurementId, fileContent);
             return Ok();
         }
 
-        //temporary here. consider move to sdLitica.Helpers
+        //temporarily here. consider move to sdLitica.Helpers
         public static async Task<List<string>> ReadAsStringAsync(IFormFile file)
         {
             List<string> result = new List<string>();
@@ -79,6 +98,10 @@ namespace sdLitica.WebAPI.Controllers.v1
             return result;
         }
 
+        /// <summary>
+        /// This REST API handler returns all time-series metadata objects owned by user
+        /// </summary>
+        /// <returns></returns>
         [HttpGet]
         [Route("temp")]
         public IActionResult GetTimeSeriesMetadataByUser()
@@ -102,6 +125,11 @@ namespace sdLitica.WebAPI.Controllers.v1
         //TODO: add content
         public IActionResult UpdateTimeSeriesMetadata([FromBody] TimeSeriesMetadataModel model)
         {
+            TimeSeriesMetadata timeSeriesMetadata = _timeSeriesMetadataService.GetTimeSeriesMetadata(model.Id);
+            if (timeSeriesMetadata == null || !timeSeriesMetadata.UserId.Equals(UserId))
+            {
+                return NotFound("this user does not have this time-series");
+            } 
             _timeSeriesMetadataService.UpdateTimeSeriesMetadata(model.Id, model.Name, model.Description).Wait();
             return Ok("ok");
         }
@@ -157,11 +185,32 @@ namespace sdLitica.WebAPI.Controllers.v1
         }
 
         /// <summary>
+        /// This REST API handler returns time-series metadata given by timeSeriesMetadataId
+        /// </summary>
+        /// <param name="timeSeriesMetadataId"></param>
+        /// <returns>
+        ///    200 - Time-series metadata
+        ///    404 - If time series doesn't exists or it is not accessible by current user
+        /// </returns>
+        [HttpGet]
+        [Route("{timeSeriesMetadataId}")]
+        public IActionResult GetTimeSeriesMetadataById([FromRoute] string timeSeriesMetadataId)
+        {
+            TimeSeriesMetadata timeSeriesMetadata = _timeSeriesMetadataService.GetTimeSeriesMetadata(timeSeriesMetadataId);
+            if (timeSeriesMetadata == null || !timeSeriesMetadata.UserId.Equals(UserId))
+            {
+                return NotFound("this user does not have this time-series");
+            }
+            return Ok(new TimeSeriesMetadataModel(timeSeriesMetadata));
+        }
+
+        /// <summary>
         /// This REST API handler returns timeseries data by id
         /// </summary>
         /// <returns>Timeseries data, instead - 404</returns>
         [HttpGet]
         [Route("{timeseriesId}/data")]
+        [AllowAnonymous]
         public IActionResult GetTimeSeriesDataById(string timeseriesId, int pageSize = 20, int offset = 0)
         {
             var measurementsResult = _timeSeriesService.ReadMeasurementById(timeseriesId).Result;
@@ -210,6 +259,7 @@ namespace sdLitica.WebAPI.Controllers.v1
         /// </summary>
         /// <returns>List of timeseries</returns>
         [HttpGet]
+        [AllowAnonymous]
         public IActionResult GetAllTimeSeries(int pageSize = 20, int offset = 0)
         {
             var measurementsResult = _timeSeriesService.ReadAllMeasurements().Result;
@@ -245,7 +295,13 @@ namespace sdLitica.WebAPI.Controllers.v1
         [Route("{timeseriesId}")]
         public IActionResult DeleteTimeSeriesById(string timeSeriesMetadataId)
         {
-            string timeSeriesId = _timeSeriesMetadataService.GetTimeSeriesMetadata(timeSeriesMetadataId).InfluxId.ToString();
+            TimeSeriesMetadata timeSeriesMetadata = _timeSeriesMetadataService.GetTimeSeriesMetadata(timeSeriesMetadataId);
+            if (timeSeriesMetadata == null || !timeSeriesMetadata.UserId.Equals(UserId))
+            {
+                return NotFound("user does not have this time-series");
+            }
+
+            string timeSeriesId = timeSeriesMetadata.InfluxId.ToString();
             var result = _timeSeriesService.DeleteMeasurementById(timeSeriesId).Result;
             if (result.Succeeded)
             {
